Move auxiliary (non-fix-related) functions out of fix module
