+++ Lib/gftools/utils.py
@@ -239,3 +239,92 @@ def mkdir(path, overwrite=True):
         os.makedirs(path)
     return path
 
+
+## Font-related utility functions
+
+def font_stylename(ttFont):
+    """Get a font's stylename using the name table. Since our fonts use the
+    RIBBI naming model, use the Typographic SubFamily Name (NAmeID 17) if it
+    exists, otherwise use the SubFamily Name (NameID 2).
+
+    Args:
+        ttFont: a TTFont instance
+    """
+    return get_name_record(ttFont, 17, fallbackID=2)
+
+
+def font_familyname(ttFont):
+    """Get a font's familyname using the name table. since our fonts use the
+    RIBBI naming model, use the Typographic Family Name (NameID 16) if it
+    exists, otherwise use the Family Name (Name ID 1).
+
+    Args:
+        ttFont: a TTFont instance
+    """
+    return get_name_record(ttFont, 16, fallbackID=1)
+
+
+def get_name_record(ttFont, nameID, fallbackID=None, platform=(3, 1, 0x409)):
+    """Return a name table record which has the specified nameID.
+
+    Args:
+        ttFont: a TTFont instance
+        nameID: nameID of name record to return,
+        fallbackID: if nameID doesn't exist, use this nameID instead
+        platform: Platform of name record. Default is Win US English
+
+    Returns:
+        str
+    """
+    name = ttFont["name"]
+    record = name.getName(nameID, 3, 1, 0x409)
+    if not record and fallbackID:
+        record = name.getName(fallbackID, 3, 1, 0x409)
+    if not record:
+        raise ValueError(f"Cannot find record with nameID {nameID}")
+    return record.toUnicode()
+
+
+def family_bounding_box(ttFonts):
+    y_min = min(f["head"].yMin for f in ttFonts)
+    y_max = max(f["head"].yMax for f in ttFonts)
+    return y_min, y_max
+
+
+def typo_metrics_enabled(ttFont):
+    return True if ttFont["OS/2"].fsSelection & (1 << 7) else False
+
+
+def family_is_vf(ttFonts):
+    has_fvar = ["fvar" in ttFont for ttFont in ttFonts]
+    if any(has_fvar):
+        if all(has_fvar):
+            return True
+        raise ValueError("Families cannot contain both static and variable fonts")
+    return False
+
+
+def validate_family(ttFonts):
+    family_is_vf(ttFonts)
+    family_names = set(font_familyname(f) for f in ttFonts)
+    if len(family_names) != 1:
+        raise ValueError(f"Multiple families found {family_names}")
+    return True
+
+
+def unique_name(ttFont, nameids):
+    font_version = _font_version(ttFont)
+    vendor = ttFont["OS/2"].achVendID.strip()
+    ps_name = nameids[6]
+    return f"{font_version};{vendor};{ps_name}"
+
+
+def _font_version(font, platEncLang=(3, 1, 0x409)):
+    nameRecord = font["name"].getName(5, *platEncLang)
+    if nameRecord is None:
+        return f'{font["head"].fontRevision:.3f}'
+    # "Version 1.101; ttfautohint (v1.8.1.43-b0c9)" --> "1.101"
+    # Also works fine with inputs "Version 1.101" or "1.101" etc
+    versionNumber = nameRecord.toUnicode().split(";")[0]
+    return versionNumber.lstrip("Version ").strip()
+
